Split core semantic gate tests
