+++ src/typechecker/tests/core_semantics.rs
@@ -3,7 +3,8 @@ use super::*;
 mod binary_ops;
 mod calls_and_returns;
 mod enum_assignment_and_modules;
-mod literals;
+mod feature_gates;
+mod intrinsic_gates;
 mod match_semantics;
 mod struct_literals;
 mod type_helpers;
